backend/requirements.txt
# Core Django
Django==5.2.12
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.2

# Storage & Media
django-cloudinary-storage==0.3.0
cloudinary==1.44.1

# Database
dj-database-url~=2.3.0
psycopg2-binary~=2.9.10

# Static Files
whitenoise~=6.9.0

# Real-time & WebSockets (ASGI)
daphne==4.2.1
channels==4.3.2

# Authentication & OAuth
google-auth==2.49.1
social-auth-app-django==5.7.0

# Production Utilities
python-dotenv==1.2.2
gunicorn~=23.0.0
